feat : Menu CRUD 기능 추가 및 Store 기능에 오너 인증 로직 추가

+++ src/main/java/com/team2/grabtable/domain/store/service/StoreServiceImpl.java
@@ -56,25 +56,28 @@ public class StoreServiceImpl implements StoreService {
     }
 
     @Override
-    public StoreResultDto getStoreDetail(Long storeId) {
+    public StoreResultDto getStoreDetail(OwnerDetails ownerDetails, Long storeId) {
         StoreResultDto storeResultDto = new StoreResultDto();
 
         try {
-            Optional<Store> storeList = storeRepository.findById(storeId);
+            Optional<Store> optionalStore = storeRepository.findById(storeId);
 
-            if (storeList.isPresent()) {
-                Store store = storeList.get();
+            if (optionalStore.isPresent()) {
+                Store store = optionalStore.get();
+                if (store.getOwner().getOwnerId().equals(ownerDetails.getOwner().getOwnerId())) {
+                    StoreDto storeDto = StoreDto.builder()
+                            .storeId(store.getStoreId())
+                            .ownerId(store.getOwner().getOwnerId())
+                            .name(store.getName())
+                            .location(store.getLocation())
+                            .type(store.getType())
+                            .build();
 
-                StoreDto storeDto = StoreDto.builder()
-                        .storeId(store.getStoreId())
-                        .ownerId(store.getOwner().getOwnerId())
-                        .name(store.getName())
-                        .location(store.getLocation())
-                        .type(store.getType())
-                        .build();
-
-                storeResultDto.setStoreDto(storeDto);
-                storeResultDto.setResult("success");
+                    storeResultDto.setStoreDto(storeDto);
+                    storeResultDto.setResult("success");
+                } else {
+                    storeResultDto.setResult("no permission at " + storeId + " store");
+                }
             } else {
                 storeResultDto.setResult("notfound");
             }
@@ -88,7 +91,7 @@ public class StoreServiceImpl implements StoreService {
     }
 
     @Override
-    public StoreImageDto getStoreImage(Long storeId) {
+    public StoreImageDto getStoreImage(OwnerDetails ownerDetails, Long storeId) {
         Store store = storeRepository.findById(storeId)
                 .orElseThrow(() -> new EntityNotFoundException("Store not found"));
         return new StoreImageDto(store.getImage(), store.getImageContentType());
@@ -122,7 +125,7 @@ public class StoreServiceImpl implements StoreService {
     public StoreResultDto updateStore(OwnerDetails ownerDetails, Long storeId, StoreRegisterDto storeRegisterDto) throws IOException {
         StoreResultDto storeResultDto = new StoreResultDto();
 
-        Store store = Store.builder()
+        Store updateStore = Store.builder()
                 .storeId(storeId)
                 .owner(ownerDetails.getOwner())
                 .name(storeRegisterDto.getName())
@@ -133,8 +136,21 @@ public class StoreServiceImpl implements StoreService {
                 .build();
 
         try {
-            storeRepository.save(store);
-            storeResultDto.setResult("success");
+
+            Optional<Store> store = storeRepository.findById(storeId);
+
+            if (store.isPresent()) {
+                Store storeToDelete = store.get();
+                if (storeToDelete.getOwner().getOwnerId().equals(ownerDetails.getOwner().getOwnerId())) {
+                    storeRepository.save(updateStore);
+                    storeResultDto.setResult("success");
+                } else {
+                    storeResultDto.setResult("no permission at " + storeId + "store");
+                }
+            } else {
+                storeResultDto.setResult("notfound");
+            }
+
         } catch (Exception e) {
             e.printStackTrace();
             storeResultDto.setResult("fail");
@@ -143,15 +159,26 @@ public class StoreServiceImpl implements StoreService {
         return storeResultDto;
     }
 
-    // todo : 사진 수정
-
     @Override
-    public StoreResultDto deleteStore(Long storeId) {
+    public StoreResultDto deleteStore(OwnerDetails ownerDetails, Long storeId) {
         StoreResultDto storeResultDto = new StoreResultDto();
 
         try {
-            storeRepository.deleteById(storeId);
-            storeResultDto.setResult("success");
+
+            Optional<Store> store = storeRepository.findById(storeId);
+
+            if (store.isPresent()) {
+                Store storeToDelete = store.get();
+                if (storeToDelete.getOwner().getOwnerId().equals(ownerDetails.getOwner().getOwnerId())) {
+                    storeRepository.deleteById(storeId);
+                    storeResultDto.setResult("success");
+                } else {
+                    storeResultDto.setResult("no permission at " + storeId + "store");
+                }
+            } else {
+                storeResultDto.setResult("notfound");
+            }
+
         } catch (Exception e) {
             e.printStackTrace();
             storeResultDto.setResult("fail");
